+++ src/Activity/Type/ProjetCreatedActivity.php
@@ -39,7 +39,7 @@ class ProjetCreatedActivity implements ActivityInterface
         $resolver->setAllowedTypes('createdBy', 'integer');
     }
 
-    public function render(array $activityParameters, string $activityType): string
+    public function render(array $activityParameters, Activity $activity): string
     {
         return sprintf(
             '%s %s a créé le projet %s.',