+++ src/Entity/Projet.php
@@ -10,6 +10,7 @@ use App\Validator as AppAssert;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
+use Symfony\Component\Serializer\Annotation\Groups;
 use Symfony\Component\Validator\Constraints as Assert;
 
 /**
@@ -25,6 +26,8 @@ class Projet implements HasSocieteInterface
      * @ORM\Id()
      * @ORM\GeneratedValue()
      * @ORM\Column(type="integer")
+     *
+     * @Groups({"recentProjets"})
      */
     private $id;
 
@@ -91,6 +94,8 @@ class Projet implements HasSocieteInterface
 
     /**
      * @ORM\Column(type="string", length=255)
+     *
+     * @Groups({"recentProjets"})
      */
     private $acronyme;